Etc: 코드 컨벤션 수정, KeyChainManager -> KeyChainService로 네이밍 변경
- 파라미터, 옵셔널 언래핑 줄바꿈 수정
